requirements.txt
astroid==2.5.6
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
DateTime==4.3
Flask==2.0.1
idna==2.10
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mvg-api==1.2.7
pylint==2.8.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zope.interface==5.4.0
